Add client layout for buyers panel
