+++ app/(buyers)/layout.js
@@ -1,42 +1,22 @@
-"use client";
+import BuyersLayoutClient from "@/components/BuyerPanel/BuyersLayoutClient";
+import { dbConnect } from "@/lib/dbConnect";
+import Category from "@/model/Category";
 
-import Header from "@/components/BuyerPanel/Header.jsx";
-import NavigationBar from "@/components/BuyerPanel/NavigationBar";
-import Footer from "@/components/BuyerPanel/Footer.jsx";
-import { usePathname } from "next/navigation";
-import { useState, useEffect } from "react";
+export default async function BuyersLayout({ children }) {
+  const categories = await getCategories();
+  return <BuyersLayoutClient categories={categories}>{children}</BuyersLayoutClient>;
+}
 
-export default function BuyersPanelLayout({ children }) {
-  const pathname = usePathname();
-  const showFooter = pathname === "/home" || pathname === "/cart";
-  const [isMenuOpen, setIsMenuOpen] = useState(true);
-
-  useEffect(() => {
-    setIsMenuOpen(true);
-  }, []);
-
-
-  return (
-    <div className="relative">
-      {/* Navbar fixed over HeroSection */}
-      <Header
-        onMenuToggle={() => setIsMenuOpen((prev) => !prev)}
-        isMenuOpen={isMenuOpen}
-      />
-
-      {/* Category Navigation */}
-      <NavigationBar
-        isMenuOpen={isMenuOpen}
-        onMenuClose={() => setIsMenuOpen(false)}
-      />
-
-      {/* Main Content */}
-      <main className="min-h-[calc(100vh-68px)] hide-scrollbar">
-        {children}
-      </main>
-
-      {/* Footer only on specific pages */}
-      {showFooter && <Footer />}
-    </div>
-  );
+async function getCategories() {
+  try {
+    await dbConnect();
+    const categories = await Category.find({ published: true })
+      .sort({ sortOrder: 1 })
+      .select("name slug subCategories")
+      .lean();
+    return categories;
+  } catch (error) {
+    console.error("Failed to load categories", error);
+    return [];
+  }
 }
